+++ core-ng/src/main/java/core/framework/impl/redis/RedisImpl.java
@@ -64,7 +64,7 @@ public final class RedisImpl implements Redis {
     }
 
     private BinaryJedis createClient() {
-        if (host == null) throw new Error("host must not be null");
+        if (host == null) throw new Error("redis.host must not be null, please check config");
         BinaryJedis client = new BinaryJedis(host, Protocol.DEFAULT_PORT, (int) timeout.toMillis());
         client.connect();
         return client;